+++ assets/js/fiber-image.js
@@ -5,4 +5,12 @@ jQuery(document).ready(function($){
     $('body').on('contextmenu', 'img', function(e){
         return false;
     });
+
+    /**
+     * Disable drag image into html page v1.1.2
+     */
+
+    $('img').on('dragstart', function(event){
+        event.preventDefault();
+    });
 });