+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.1.1 - 2021-09-22
+
+* Fixed: Option Disable Right Click Image is not working
+* Changed: Rename option Disable Right Click Image with Disable Image Protection
+
 ## 1.1.1 - 2021-09-21
 
 * Fixed: Issue with custom taxonomy when enabling Custom Post Order